requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-cloud-vision>=3.6.0
pillow>=10.2.0
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'
structlog>=24.1.0
slowapi>=0.1.9
